Updated New Soluations About TODO List But Mistake How to relation task with task list

+++ partical_assignments/PA6_NadeenDames/TODO/user.py
@@ -8,20 +8,25 @@ bp = Blueprint('user', __name__)
 
 @bp.route('/add/user', methods=['GET', 'POST'])
 def add_user():
-    
+    if "uid" not in session:
+            return redirect('/login')
+
     if request.method == 'GET':
         # render add user blueprint
-        return render_template('user/index.html')
+        return render_template('login/register.html')
     else:
         username = request.form['username']
         password = request.form['password']
+        email = request.form['email']
+        firstname = request.form['firstname']
+        lastname = request.form['lastname']
 
         # get the DB connection
         db = get_db()
 
         # insert user into db
         try:
-            db.execute("INSERT INTO user (username, password) VALUES (?, ?);", (username, password))
+            db.execute("INSERT INTO user (username,fname,lname, password,email) VALUES (?, ?,?,?,?);", (username,firstname,lastname,password,email))
             db.commit()
             
             return redirect("/users")
@@ -32,6 +37,8 @@ def add_user():
 
 @bp.route('/users')
 def get_users():
+    if "uid" not in session:
+            return redirect('/login')
     # get the DB connection
     db = get_db()
 
@@ -41,78 +48,38 @@ def get_users():
     # render 'list.html' blueprint with users
     return render_template('user/list.html', users=users)
 
-@bp.route('/delete')
-def delete():
-      
-        db = get_db()
-        # insert user into db
-        try:
-            db.execute('DELETE FROM user')
-            db.commit()
-            
-            return render_template("login/login.html")
-
-        except sqlite3.Error as er:
-            print('SQLite error: %s' % (' '.join(er.args)))
-            return redirect("/404")
-@bp.route('/user/update')
-def update_username():
-    if request.method == 'GET':
-        # render add user blueprint
-        return render_template('user/update-username.html')
+@bp.route('/users/updateinfo',methods=['POST','GET'])
+def updateinfo():
+    if "uid" not in session:
+            return redirect('/login')
+    if request.method=='GET':
+        return render_template('user/updateinfo.html')
     else:
-        username = request.form['username_update']
-        # get the DB connection
-        db = get_db()
+        db=get_db()
+        username=request.form['username']
+        firstname=request.form['firstname']
+        lastname=request.form['lastname']
+        email=request.form['email']
+        password=request.form['password']
 
-        # insert user into db
         try:
-            db.execute("UPDATE  user SET username=?",(username))
+            db.execute("UPDATE user set username=?,fname=?,lname=?,email=?,password=?;",(username,firstname,lastname,email,password))
             db.commit()
-
-            return redirect("/users")
-
-        except sqlite3.Error as er:
-            print('SQLite error: %s' % (' '.join(er.args)))
-            return redirect("/404")
-@bp.route('/user/update/email')
-def update_email():
-    if request.method == 'GET':
-        # render add user blueprint
-        return render_template('user/update-email.html')
-    else:
-        email = request.form['email_update']
-        # get the DB connection
-        db = get_db()
-
-        # insert user into db
-        try:
-            db.execute("UPDATE  user SET email=?",(email))
-            db.commit()
-            
-            return redirect("/users")
-
+            return redirect('/users')
         except sqlite3.Error as er:
             print('SQLite error: %s' % (' '.join(er.args)))
             return redirect("/404")
 
-@bp.route('/user/update')
-def update_password():
-    if request.method == 'GET':
-        # render add user blueprint
-        return render_template('user/update-password.html')
-    else:
-        password = request.form['password_update']
-        # get the DB connection
-        db = get_db()
-
-        # insert user into db
+@bp.route('/users/deleteuser')
+def deleteaccount():
+        if "uid" not in session:
+            return redirect('/login')
+        db=get_db()
         try:
-            db.execute("UPDATE  user SET password=?",(password))
+            db.execute("DELETE FROM user")
             db.commit()
-            
-            return redirect("/users")
-
+            return render_template('login/register.html')
         except sqlite3.Error as er:
             print('SQLite error: %s' % (' '.join(er.args)))
             return redirect("/404")
+
